pyproject.toml
[project]
name = "valheim-for-the-boys"
version = "0.1.0"
description = "Helpers for installing Valheim mods for a dedicated server"
authors = [ { name = "", email = "" } ]
readme = "README.md"
requires-python = ">=3.9"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.poetry.dev-dependencies]
pyyaml = "^6.0"
